Update Deity Anguana Behavior for Deity Battle
Now Deity Anguana at each turn attacks the Player with Spiteful Attack move.

+++ Assets/Code/Scripts/Unit.cs
@@ -86,7 +86,7 @@ public class Unit : MonoBehaviour
         }
     }
 
-    public void TakeDamage(int receivedDamage)
+    public void TakeDamage(float receivedDamage)
     {
         HealthPoints -= receivedDamage;
         OnTakenDamage.Invoke(receivedDamage);
